pyproject.toml
[project]
name = "insurance_calc"
version = "0.1.0"
description = "A simple insurance calculation application"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.112.2",
    "alembic>=1.13.2",
    "uvicorn>=0.30.6",
    "gunicorn>=23.0.0",
    "pydantic-settings>=2.2.1",
    "yarl>=1.9.4",
    "passlib>=1.7.4",
    "python-jose>=3.3.0",
    "redis>=5.0.8",
    "typer>=0.12.5",
    "loguru>=0.7.2",
    "orjson>=3.10.7",
    "uvloop>=0.20.0",
    "httptools>=0.6.1",
    "sqlalchemy>=2.0.36",
    "aiofiles>=24.1.0",
    "asyncpg>=0.30.0",
    "aiokafka>=0.12.0",
]

[tool.uv]
dev-dependencies = [
    "fakeredis>=2.26.1",
    "httpx>=0.27.2",
    "pytest>=8.3.3",
    "ruff>=0.6.2",
]

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
    "B904", # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
